feat: add notifications
