+++ Workshop-Session-and-Authentication/static/css/site.css
@@ -281,3 +281,16 @@ footer {
         opacity: 0;
     }
 }
+#notifications {
+    margin: 0 auto;
+    margin-top: 0px;
+    margin-bottom: 0px;
+    text-align: center;
+    width: 50%;
+    padding: 0.5%;
+    margin-top: 2%;
+    margin-bottom: 2%;
+    background: rgb(231, 118, 118);
+    color: white;
+    border-radius: 5px;
+}
