+++ Workshop-Session-and-Authentication/views/layouts/main.hbs
@@ -11,6 +11,7 @@
 <body>
   <div class="container">
     {{> header}}
+    {{> notification}}
     <main>
       {{{body}}}
     </main>
