+++ Workshop-Session-and-Authentication/views/partials/notification.hbs
@@ -1,0 +1,5 @@
+{{#if error}}
+<div id="notifications">
+    <p id="notification-message">{{error}}</p>
+</div>
+{{/if}}
